Its actually doable in raw js using JSDoc
